@@ -1,30 +1,19 @@
 <template>
   <div class="h-screen w-full overflow-hidden select-none relative">
     <canvas ref="canvasEl" style="cursor: crosshair"></canvas>
-    <Tools :canvas="canvas" />
+    <Tools :canvas="canvas!" />
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent, h, nextTick, onMounted, ref, warn, watch } from "vue";
-import Canvas from "./utils/Canvas";
-import Tools from "./components/Tools.vue";
+<script lang="ts" setup>
+import { onMounted, ref, useTemplateRef } from 'vue'
+import Canvas from './utils/Canvas'
+import Tools from './components/Tools.vue'
 
-export default defineComponent({
-  name: "App",
-  components: {
-    Tools,
-  },
-  setup() {
-    const canvasEl = ref<HTMLCanvasElement | null>(null);
-    const canvas = ref<Canvas | null>(null);
-    onMounted(() => {
-      canvas.value = new Canvas(canvasEl.value!);
-    });
-    return {
-      canvasEl,
-      canvas,
-    };
-  },
-});
+const canvasEl = useTemplateRef<HTMLCanvasElement>('canvasEl')
+const canvas = ref<Canvas>()
+
+onMounted(() => {
+  canvas.value = new Canvas(canvasEl.value!)
+})
 </script>
