@@ -9,7 +9,7 @@
             @click="onPaint"
             title="画笔"
           >
-            <Write theme="outline" size="36" :fill="color" />
+            <PencilLine :fill="color" />
           </div>
         </template>
         <div class="pt-4 w-80 px-2">
@@ -22,10 +22,7 @@
               @change="onWidthChange"
             ></el-slider>
             <div class="w-8 ml-10 flex justify-center items-center">
-              <div
-                class="rounded-full border border-gray-300"
-                :style="style"
-              ></div>
+              <div class="rounded-full border border-gray-300" :style="style"></div>
             </div>
           </div>
           <div class="flex justify-around items-center">
@@ -47,142 +44,104 @@
         @click="onRemove"
         title="擦除"
       >
-        <ClearFormat theme="outline" size="36" fill="#333" />
+        <Eraser />
       </div>
       <div
         class="rounded-full cursor-pointer border border-gray-200 p-2 mx-1"
         @click="onSave"
         title="保存"
       >
-        <Save theme="outline" size="36" fill="#333" />
+        <Save />
       </div>
       <div
         class="rounded-full cursor-pointer border border-gray-200 p-2 mx-1 history"
         @click="onUndo"
         title="撤回"
       >
-        <Undo theme="outline" size="36" fill="#333" />
+        <Undo />
       </div>
       <div
         class="rounded-full cursor-pointer border border-gray-200 p-2 mx-1 history"
         @click="onRedo"
         title="恢复"
       >
-        <Redo theme="outline" size="36" fill="#333" />
+        <Redo />
       </div>
       <div
         class="rounded-full cursor-pointer border border-gray-200 p-2 mx-1"
         @click="onClear"
         title="清空画布"
       >
-        <Clear theme="outline" size="36" fill="#333" />
+        <CircleX />
       </div>
     </div>
   </section>
 </template>
 
-<script lang="ts">
-import { computed, defineComponent, ref } from "vue";
-import {
-  Write,
-  Clear,
-  ClearFormat,
-  Save,
-  Undo,
-  Redo,
-} from "@icon-park/vue-next";
-import Canvas from "../utils/Canvas";
-import { ElMessageBox } from "element-plus";
+<script lang="ts" setup>
+import { computed, ref } from 'vue'
+import Canvas from '../utils/Canvas'
+import { ElMessageBox } from 'element-plus'
+import { CircleX, Redo, Save, Undo, PencilLine, Eraser } from 'lucide-vue-next'
 
-type Active = "paint" | "remove";
+type Active = 'paint' | 'remove'
 
-export default defineComponent({
-  name: "Tools",
-  props: {
-    canvas: {
-      type: Canvas,
-      required: true,
-    },
-  },
-  components: {
-    Write,
-    Clear,
-    ClearFormat,
-    Save,
-    Undo,
-    Redo,
-  },
-  setup(props) {
-    const active = ref<Active>("paint");
-    const width = ref(20);
-    const color = ref("#000000");
-    const selectColor = (c: string) => {
-      color.value = c;
-      props.canvas.setColor(c);
-    };
-    const onWidthChange = (v: number) => {
-      props.canvas.setWidth(v);
-    };
-    // 画笔
-    const onPaint = () => {
-      props.canvas.paint();
-      active.value = "paint";
-    };
-    // 清空
-    const onClear = async () => {
-      await ElMessageBox.confirm("确认清空画板?", "清空", {
-        confirmButtonText: "确定",
-        cancelButtonText: "取消",
-        type: "warning",
-      });
-      props.canvas.clear();
-    };
-    // 橡皮
-    const onRemove = () => {
-      props.canvas.remove();
-      active.value = "remove";
-    };
-    // 保存
-    const onSave = () => {
-      props.canvas.save();
-    };
-    // onUndo
-    const onUndo = () => {
-      props.canvas.undo();
-    };
-    // onUndo
-    const onRedo = () => {
-      props.canvas.redo();
-    };
-    return {
-      width,
-      color,
-      style: computed(() => ({
-        width: width.value + "px",
-        height: width.value + "px",
-        backgroundColor: color.value,
-      })),
-      colors: [
-        "#ffffff",
-        "#000000",
-        "#FF3333",
-        "#0066FF",
-        "#FFFF33",
-        "#33CC66",
-        "#808080",
-      ],
-      onWidthChange,
-      selectColor,
-      onClear,
-      onRemove,
-      onPaint,
-      onSave,
-      onUndo,
-      onRedo,
-      active,
-    };
-  },
-});
+const props = defineProps({
+  canvas: { type: Canvas, required: true },
+})
+
+const active = ref<Active>('paint')
+const width = ref(20)
+const color = ref('#000000')
+const selectColor = (c: string) => {
+  color.value = c
+  props.canvas.setColor(c)
+}
+const onWidthChange = (v: number) => {
+  props.canvas.setWidth(v)
+}
+// 画笔
+const onPaint = () => {
+  props.canvas.paint()
+  active.value = 'paint'
+}
+// 清空
+const onClear = async () => {
+  ElMessageBox.confirm('确认清空画板?', '清空', {
+    confirmButtonText: '确定',
+    cancelButtonText: '取消',
+    type: 'warning',
+  })
+    .then(() => {
+      props.canvas.clear()
+    })
+    .catch(() => {})
+}
+// 橡皮
+const onRemove = () => {
+  props.canvas.remove()
+  active.value = 'remove'
+}
+// 保存
+const onSave = () => {
+  props.canvas.save()
+}
+// onUndo
+const onUndo = () => {
+  props.canvas.undo()
+}
+// onUndo
+const onRedo = () => {
+  props.canvas.redo()
+}
+
+const style = computed(() => ({
+  width: width.value + 'px',
+  height: width.value + 'px',
+  backgroundColor: color.value,
+}))
+
+const colors = ['#ffffff', '#000000', '#FF3333', '#0066FF', '#FFFF33', '#33CC66', '#808080']
 </script>
 
 <style scoped>
