@@ -1,10 +1,10 @@
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
 import { createApp } from 'vue'
-// TypeScript error? Run VSCode command
-// TypeScript: Select TypeScript version - > Use Workspace Version
 import App from './App.vue'
 import './index.css'
-import Element from './plugins/element'
 
-createApp(App)
-  .use(Element)
-  .mount('#app')
+const app = createApp(App)
+
+app.use(ElementPlus)
+app.mount('#app')
